refactors and adjusts the calendar view UI
