+++ washee_app/lib/core/pages/home_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:washee/features/unlock/presentation/pages/wash_screen.dart';
 import 'package:washee/core/pages/washee_screen.dart';
-import 'package:washee/core/pages/calendar_screen.dart';
+import 'package:washee/features/booking/presentation/pages/calendar_screen.dart';
 
 class HomeScreen extends StatefulWidget {
   static const routeName = "/home-screen";
