+++ washee_app/lib/main.dart
@@ -3,6 +3,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:provider/provider.dart';
 import 'package:washee/core/pages/home_screen.dart';
 import 'package:washee/core/providers/global_provider.dart';
+import 'package:washee/features/booking/presentation/provider/calendar_provider.dart';
 import 'package:washee/features/unlock/presentation/provider/unlock_provider.dart';
 import 'injection_container.dart' as ic;
 import 'core/presentation/themes/themes.dart';
@@ -20,6 +21,7 @@ class WasheeApp extends StatelessWidget {
       providers: [
         ChangeNotifierProvider(create: (ctx) => UnlockProvider()),
         ChangeNotifierProvider(create: (ctx) => GlobalProvider()),
+        ChangeNotifierProvider(create: (ctx) => CalendarProvider()),
       ],
       child: ScreenUtilInit(
         designSize: Size(1000, 1600),
